Maint: Structural changes to project, moved class parsing methods to utils/
