+++ matio/utils/__init__.py
@@ -1,0 +1,6 @@
+# matio/utils/__init__.py
+from .matstring import toString
+from .mattables import toDataFrame
+from .mattimes import toDatetime, toDuration
+
+__all__ = ["toDataFrame", "toDatetime", "toDuration", "toString"]
